+++ .env.example
@@ -1,0 +1,4 @@
+PASSWORD=""
+DATABASE_URL="postgresql://username:PASSWORD@localhost:5432/database?schema=public"
+JWT_SECRET=""
+JWT_EXP_TIME=""
