+++ .gitignore
@@ -40,5 +40,5 @@ db.sqlite3
 dev.db
 **.db.**
 
-# environments
+# environment
 .env
