+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "procuraqui-api",
+  "name": "bknd-jwt",
   "version": "0.0.1",
   "description": "",
   "author": "",
@@ -26,10 +26,10 @@
     "@nestjs/mapped-types": "*",
     "@nestjs/passport": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
-    "@prisma/client": "4.15.0",
+    "@prisma/client": "5.0.0",
     "bcrypt": "^5.1.0",
-    "class-transformer": "^0.5.1",
-    "class-validator": "^0.14.0",
+    "class-transformer": "^0.2.3",
+    "class-validator": "^0.11.1",
     "passport": "^0.6.0",
     "passport-jwt": "^4.0.1",
     "passport-local": "^1.0.0",
@@ -44,7 +44,7 @@
     "@types/express": "^4.17.17",
     "@types/jest": "^29.5.2",
     "@types/node": "^20.3.1",
-    "@types/passport-jwt": "^3.0.8",
+    "@types/passport-jwt": "^3.0.9",
     "@types/passport-local": "^1.0.35",
     "@types/supertest": "^2.0.12",
     "@typescript-eslint/eslint-plugin": "^5.59.11",
@@ -54,7 +54,7 @@
     "eslint-plugin-prettier": "^4.2.1",
     "jest": "^29.5.0",
     "prettier": "^2.8.8",
-    "prisma": "^4.15.0",
+    "prisma": "^5.0.0",
     "source-map-support": "^0.5.21",
     "supertest": "^6.3.3",
     "ts-jest": "^29.1.0",
