+++ src/app.module.ts
@@ -1,21 +1,21 @@
 import { Module } from '@nestjs/common'
+import { APP_GUARD } from '@nestjs/core'
 import { AppController } from './app.controller'
 import { AppService } from './app.service'
-import { AuthModule } from './auth/auth.module'
-import { PrismaModule } from './prisma/prisma.module'
-import { UserModule } from './user/user.module'
-import { APP_GUARD } from '@nestjs/core'
-import { JwtAuthGuard } from './auth/guards/jwt-auth.guard'
+import { AuthModule } from './app/auth/auth.module'
+import { JwtAuthGuard } from './app/auth/guards/jwt-auth.guard'
+import { PrismaModule } from './db/prisma/prisma.module'
+import { UserModule } from './app/user/user.module'
 
 @Module({
-    imports: [AuthModule, PrismaModule, UserModule],
-    controllers: [AppController],
-    providers: [
-        AppService,
-        {
-            provide: APP_GUARD,
-            useClass: JwtAuthGuard
-        }
-    ]
+	imports: [PrismaModule, UserModule, AuthModule],
+	controllers: [AppController],
+	providers: [
+		AppService,
+		{
+			provide: APP_GUARD,
+			useClass: JwtAuthGuard,
+		},
+	],
 })
 export class AppModule {}
