+++ src/app.service.ts
@@ -1,17 +1,8 @@
-import { Injectable } from '@nestjs/common'
+import { Injectable } from '@nestjs/common';
 
 @Injectable()
 export class AppService {
-    getInfos(): any {
-        return {
-            message:
-                'To see more infos, show the repo https:github.com/meiazero/procuraqui-api'
-        }
-    }
-
-    getHello(): any {
-        return {
-            message: 'Hello World!'
-        }
-    }
+  getHello(): string {
+    return 'Hello World!';
+  }
 }
