+++ src/app/auth/auth.module.ts
@@ -1,0 +1,27 @@
+import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common'
+import { JwtModule } from '@nestjs/jwt'
+import { PassportModule } from '@nestjs/passport'
+import { UserModule } from '../user/user.module'
+import { AuthController } from './auth.controller'
+import { AuthService } from './auth.service'
+import { JwtStrategy } from './strategies/jwt.strategy'
+import { LocalStrategy } from './strategies/local.strategy'
+import { LoginValidationMiddleware } from './middlewares/login-validation.middleware'
+
+@Module({
+	imports: [
+		UserModule,
+		PassportModule,
+		JwtModule.register({
+			secret: process.env.JWT_SECRET,
+			signOptions: { expiresIn: '30d' },
+		}),
+	],
+	controllers: [AuthController],
+	providers: [AuthService, LocalStrategy, JwtStrategy],
+})
+export class AuthModule implements NestModule {
+	configure(consumer: MiddlewareConsumer) {
+		consumer.apply(LoginValidationMiddleware).forRoutes('login')
+	}
+}
