+++ src/app/auth/auth.service.ts
@@ -1,0 +1,47 @@
+import { Injectable } from '@nestjs/common'
+import { JwtService } from '@nestjs/jwt'
+import { compare } from 'bcrypt'
+import { UnauthorizedError } from './errors/unauthorized.error'
+import { User } from '../user/entities/user.entity'
+import { UserService } from '../user/user.service'
+import { UserPayload } from './models/UserPayload'
+import { UserToken } from './models/UserToken'
+
+@Injectable()
+export class AuthService {
+	constructor(
+		private readonly jwtService: JwtService,
+		private readonly userService: UserService,
+	) {}
+
+	async login(user: User): Promise<UserToken> {
+		const payload: UserPayload = {
+			sub: user.id,
+			email: user.email,
+			name: user.name,
+		}
+
+		return {
+			access_token: this.jwtService.sign(payload),
+		}
+	}
+
+	async validateUser(email: string, password: string): Promise<User> {
+		const user = await this.userService.findUserByEmail(email)
+
+		if (user) {
+			const isPasswordValid = await compare(password, user.password)
+
+			if (isPasswordValid) {
+				return {
+					...user,
+					password: undefined,
+				}
+			}
+		}
+
+		throw new UnauthorizedError(
+			'Email address or password provided is incorrect.',
+		)
+	}
+}
