+++ src/app/auth/decorators/current-user.decorator.ts
@@ -1,0 +1,11 @@
+import { createParamDecorator, ExecutionContext } from '@nestjs/common'
+import { User } from '../../user/entities/user.entity'
+import { AuthRequest } from '../models/AuthRequest'
+
+export const CurrentUser = createParamDecorator(
+	(data: unknown, context: ExecutionContext): User => {
+		const request = context.switchToHttp().getRequest<AuthRequest>()
+
+		return request.user
+	},
+)
