+++ src/app/auth/guards/local-auth.guard.ts
@@ -1,0 +1,24 @@
+import {
+	ExecutionContext,
+	Injectable,
+	UnauthorizedException,
+} from '@nestjs/common'
+import { AuthGuard } from '@nestjs/passport'
+import { Observable } from 'rxjs'
+
+@Injectable()
+export class LocalAuthGuard extends AuthGuard('local') {
+	canActivate(
+		context: ExecutionContext,
+	): boolean | Promise<boolean> | Observable<boolean> {
+		return super.canActivate(context)
+	}
+
+	handleRequest(err, user) {
+		if (err || !user) {
+			throw new UnauthorizedException(err?.message)
+		}
+
+		return user
+	}
+}
