+++ src/app/auth/models/AuthRequest.ts
@@ -1,0 +1,6 @@
+import { Request } from 'express'
+import { User } from '../../user/entities/user.entity'
+
+export interface AuthRequest extends Request {
+	user: User
+}
