+++ src/app/auth/models/LoginRequestBody.ts
@@ -1,0 +1,9 @@
+import { IsEmail, IsString } from 'class-validator'
+
+export class LoginRequestBody {
+	@IsEmail()
+	email: string
+
+	@IsString()
+	password: string
+}
