+++ src/app/auth/models/UserFromJwt.ts
@@ -1,0 +1,5 @@
+export interface UserFromJwt {
+	id: string
+	email: string
+	name: string
+}
