+++ src/app/auth/models/UserPayload.ts
@@ -1,0 +1,7 @@
+export interface UserPayload {
+	sub: string
+	email: string
+	name: string
+	iat?: number
+	exp?: number
+}
