+++ src/app/auth/models/UserToken.ts
@@ -1,0 +1,3 @@
+export interface UserToken {
+	access_token: string
+}
