+++ src/app/auth/strategies/jwt.strategy.ts
@@ -1,0 +1,24 @@
+import { Injectable } from '@nestjs/common'
+import { PassportStrategy } from '@nestjs/passport'
+import { ExtractJwt, Strategy } from 'passport-jwt'
+import { UserFromJwt } from '../models/UserFromJwt'
+import { UserPayload } from '../models/UserPayload'
+
+@Injectable()
+export class JwtStrategy extends PassportStrategy(Strategy) {
+	constructor() {
+		super({
+			jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+			ignoreExpiration: false,
+			secretOrKey: process.env.JWT_SECRET,
+		})
+	}
+
+	async validate(payload: UserPayload): Promise<UserFromJwt> {
+		return {
+			id: payload.sub,
+			email: payload.email,
+			name: payload.name,
+		}
+	}
+}
