+++ src/app/user/entities/user.entity.ts
@@ -1,0 +1,8 @@
+export class User {
+	id?: string
+	email: string
+	name: string
+	password: string
+	createdAt?: Date
+	updateAt?: Date
+}
