+++ src/app/user/user.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common'
+import { UserController } from './user.controller'
+import { UserService } from './user.service'
+import { PrismaService } from 'src/db/prisma/prisma.service'
+
+@Module({
+	controllers: [UserController],
+	providers: [UserService, PrismaService],
+	exports: [UserService],
+})
+export class UserModule {}
