--- src/auth/auth.controller.ts
@@ -1,25 +1,0 @@
-import {
-    Controller,
-    HttpCode,
-    HttpStatus,
-    Post,
-    Request,
-    UseGuards
-} from '@nestjs/common'
-import { AuthService } from './auth.service'
-import { LocalAuthGuard } from './guards/local-auth.guard'
-import { AuthRequest } from './models/AuthRequest'
-import { IsPublic } from './decorator/is-public.decorator'
-
-@Controller()
-export class AuthController {
-    constructor(private readonly authService: AuthService) {}
-
-    @IsPublic()
-    @Post('login')
-    @HttpCode(HttpStatus.OK)
-    @UseGuards(LocalAuthGuard)
-    async login(@Request() req: AuthRequest) {
-        return this.authService.login(req.user)
-    }
-}
