--- src/auth/auth.module.ts
@@ -1,25 +1,0 @@
-import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common'
-import { AuthController } from './auth.controller'
-import { AuthService } from './auth.service'
-import { UserModule } from 'src/user/user.module'
-import { LocalStrategy } from './strategies/local.strategy'
-import { JwtModule } from '@nestjs/jwt'
-import { JwtStrategy } from './strategies/jwt.strategy'
-import { LoginValidationMiddleware } from './middleware/login-validation.middleware'
-
-@Module({
-    imports: [
-        UserModule,
-        JwtModule.register({
-            secret: process.env.JWT_SECRET,
-            signOptions: { expiresIn: process.env.EXP_TIME }
-        })
-    ],
-    controllers: [AuthController],
-    providers: [AuthService, LocalStrategy, JwtStrategy]
-})
-export class AuthModule implements NestModule {
-    configure(consumer: MiddlewareConsumer) {
-        consumer.apply(LoginValidationMiddleware).forRoutes('login')
-    }
-}
