--- src/auth/decorator/current-user.decorator.ts.ts
@@ -1,11 +1,0 @@
-import { createParamDecorator, ExecutionContext } from '@nestjs/common'
-import { User } from 'src/user/entities/user.entity'
-import { AuthRequest } from '../models/AuthRequest'
-
-export const CurrentUser = createParamDecorator(
-    (data: unknown, context: ExecutionContext): User => {
-        const request = context.switchToHttp().getRequest<AuthRequest>()
-
-        return request.user
-    }
-)
