--- src/auth/guards/jwt-auth.guard.ts
@@ -1,47 +1,0 @@
-// NestJS
-import {
-    ExecutionContext,
-    Injectable,
-    UnauthorizedException
-} from '@nestjs/common'
-import { Reflector } from '@nestjs/core'
-// Password
-import { AuthGuard } from '@nestjs/passport'
-// Decorators
-import { IS_PUBLIC_KEY } from '../decorator/is-public.decorator'
-// Error Handling
-import { UnauthorizedError } from '../erros/unauthorized.error'
-
-@Injectable()
-export class JwtAuthGuard extends AuthGuard('jwt') {
-    constructor(private reflector: Reflector) {
-        super()
-    }
-
-    canActivate(context: ExecutionContext): Promise<boolean> | boolean {
-        const isPublic = this.reflector.getAllAndOverride<boolean>(
-            IS_PUBLIC_KEY,
-            [context.getHandler(), context.getClass()]
-        )
-
-        if (isPublic) {
-            return true
-        }
-
-        const canActivate = super.canActivate(context)
-
-        if (typeof canActivate === 'boolean') {
-            return canActivate
-        }
-
-        const canActivatePromise = canActivate as Promise<boolean>
-
-        return canActivatePromise.catch((error) => {
-            if (error instanceof UnauthorizedError) {
-                throw new UnauthorizedException(error.message)
-            }
-
-            throw new UnauthorizedException()
-        })
-    }
-}
