--- src/auth/models/AuthRequest.ts
@@ -1,6 +1,0 @@
-import { Request } from 'express'
-import { User } from '../../user/entities/user.entity'
-
-export interface AuthRequest extends Request {
-    user: User
-}
