--- src/auth/models/LoginRequestBody.ts
@@ -1,9 +1,0 @@
-import { IsEmail, IsString } from 'class-validator'
-
-export class LoginRequestBody {
-    @IsEmail()
-    email: string
-
-    @IsString()
-    password: string
-}
