--- src/auth/models/UserFromJwt.ts
@@ -1,5 +1,0 @@
-export interface UserFromJwt {
-    id: string
-    email: string
-    name: string
-}
