--- src/auth/models/UserPayload.ts
@@ -1,7 +1,0 @@
-export interface UserPayload {
-    sub: string
-    email: string
-    name: string
-    iat?: number
-    exp?: number
-}
