--- src/auth/models/UserToken.ts
@@ -1,3 +1,0 @@
-export interface UserToken {
-    access_token: string
-}
