--- src/auth/strategies/local.strategy.ts
@@ -1,15 +1,0 @@
-import { Injectable } from '@nestjs/common'
-import { PassportStrategy } from '@nestjs/passport'
-import { Strategy } from 'passport-local'
-import { AuthService } from '../auth.service'
-
-@Injectable()
-export class LocalStrategy extends PassportStrategy(Strategy) {
-    constructor(private authService: AuthService) {
-        super({ usernameField: 'email' })
-    }
-
-    validate(email: string, password: string) {
-        return this.authService.validateUser(email, password)
-    }
-}
