+++ src/main.ts
@@ -1,19 +1,20 @@
+import { ValidationPipe } from '@nestjs/common'
 import { NestFactory } from '@nestjs/core'
 import { AppModule } from './app.module'
-import { ValidationPipe } from '@nestjs/common'
 
 async function bootstrap() {
-    const app = await NestFactory.create(AppModule)
+	const app = await NestFactory.create(AppModule)
 
-    // Pipes
-    app.useGlobalPipes(
-        new ValidationPipe({
-            transform: true,
-            whitelist: true,
-            forbidNonWhitelisted: true
-        })
-    )
+	// Pipes
+	app.useGlobalPipes(
+		new ValidationPipe({
+			transform: true,
+			whitelist: true,
+			forbidNonWhitelisted: true,
+		}),
+	)
 
-    await app.listen(3000)
+	await app.listen(3000)
 }
+
 bootstrap()
