--- src/user/entities/user.entity.ts
@@ -1,8 +1,0 @@
-export class User {
-    id?: string
-    name: string
-    email: string
-    password: string
-    createdAt?: Date
-    updatedAt?: Date
-}
