--- src/user/user.controller.ts
@@ -1,27 +1,0 @@
-import {
-    Controller,
-    Get,
-    Post,
-    Body,
-    Patch,
-    Param,
-    Delete
-} from '@nestjs/common'
-import { UserService } from './user.service'
-import { CreateUserDto } from './dto/create-user.dto'
-import { UpdateUserDto } from './dto/update-user.dto'
-
-@Controller('user')
-export class UserController {
-    constructor(private readonly userService: UserService) {}
-
-    @Post()
-    create(@Body() createUserDto: CreateUserDto) {
-        return this.userService.create(createUserDto)
-    }
-
-    @Get()
-    getAllUsers() {
-        return this.userService.findAll()
-    }
-}
