--- src/user/user.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common'
-import { UserService } from './user.service'
-import { UserController } from './user.controller'
-import { PrismaModule } from 'src/prisma/prisma.module'
-
-@Module({
-    imports: [PrismaModule],
-    controllers: [UserController],
-    providers: [UserService],
-    exports: [UserService]
-})
-export class UserModule {}
